+++ connector_baidu_translate/src/main/java/com/moocsk/lowcode/baidu/translate/util/MD5.java
@@ -1,4 +1,4 @@
-package com.moocsk.lowcode.util;
+package com.moocsk.lowcode.baidu.translate.util;
 
 import java.security.MessageDigest;
